added authentication for user put and delete

+++ geodata/resources/user.py
@@ -2,8 +2,9 @@ from flask import Response, request, jsonify
 from flask_restful import Resource,  url_for
 from sqlalchemy.exc import IntegrityError
 from jsonschema import validate, ValidationError, Draft7Validator
-from werkzeug.exceptions import Conflict, BadRequest, UnsupportedMediaType
+from werkzeug.exceptions import Conflict, BadRequest, UnsupportedMediaType, Forbidden
 from geodata.models import User, db
+from geodata.auth import auth, check_user_access
 
 draft7_format_checker = Draft7Validator.FORMAT_CHECKER
 
@@ -101,9 +102,13 @@ class UserItem(Resource):
         }
 
         return jsonify(response)
-
+    @auth.login_required
     def put(self, user):
         """Update user by username"""
+
+        if not check_user_access(user):
+            raise Forbidden("You don't have permission to modify this user")
+
         if request.content_type != "application/json":
             raise UnsupportedMediaType
         try:
@@ -133,9 +138,11 @@ class UserItem(Resource):
 
 
         return Response(status=204)
-
+    @auth.login_required
     def delete(self, user):
         """Delete user by username"""
+        if not check_user_access(user):
+            raise Forbidden("You don't have permission to delete this user")
         db.session.delete(user)
         db.session.commit()
         return Response(status=204)
